fix: consolidate debugLog implementations into single module
- Removed duplicate debugLog function from utilities.ts
- Updated all operations to import debugLog from debug.ts module
- Combined DEBUG and debugLog imports where both are used
- Eliminates confusion from having two different debugLog implementations
- Addresses CodeRabbitAI critical issue about duplicate functions
