+++ src/nodes/SurrealDb/resources/field/operations/deleteField.operation.ts
@@ -11,10 +11,10 @@ import {
     buildCredentialsObject,
     checkQueryResult,
 } from "../../../GenericFunctions";
-import { debugLog, addSuccessResult } from "../../../utilities";
+import { addSuccessResult  } from "../../../utilities"
 import type { IOperationHandler } from "../../../types/operation.types";
 
-import { DEBUG } from "../../../debug";
+import { DEBUG, debugLog } from "../../../debug";
 
 /**
  * Implementation of the "Delete Field" operation
