+++ src/nodes/SurrealDb/resources/record/operations/updateRecord.operation.ts
@@ -14,12 +14,11 @@ import {
 import {
     createRecordId,
     parseAndValidateRecordId,
-    debugLog,
     addSuccessResult,
     addErrorResult,
 } from "../../../utilities";
 
-import { DEBUG } from "../../../debug";
+import { DEBUG, debugLog } from "../../../debug";
 
 /**
  * Update Record operation handler for Record resource
