+++ src/nodes/SurrealDb/resources/relationship/operations/deleteRelationship.operation.ts
@@ -10,10 +10,9 @@ import {
     buildCredentialsObject,
     checkQueryResult,
 } from "../../../GenericFunctions";
-import { debugLog } from "../../../utilities";
 import type { IOperationHandler } from "../../../types/operation.types";
 
-import { DEBUG } from "../../../debug";
+import { DEBUG, debugLog } from "../../../debug";
 
 /**
  * Implementation of the "Delete Relationship" operation
