+++ src/nodes/SurrealDb/resources/table/operations/deleteAllRecords.operation.ts
@@ -7,7 +7,6 @@ import { NodeOperationError } from "n8n-workflow";
 import type { Surreal } from "surrealdb";
 import {
     formatArrayResult,
-    debugLog,
     createErrorResult,
 } from "../../../utilities";
 import {
@@ -20,7 +19,7 @@ import {
 } from "../../../GenericFunctions";
 import type { IOperationHandler } from "../../../types/operation.types";
 
-import { DEBUG } from "../../../debug";
+import { DEBUG, debugLog } from "../../../debug";
 
 /**
  * Implementation of the "Delete All Records" operation
