+++ src/nodes/SurrealDb/resources/table/operations/getMany.operation.ts
@@ -8,7 +8,6 @@ import type { Surreal } from "surrealdb";
 import {
     formatArrayResult,
     parseAndValidateRecordId,
-    debugLog,
     addErrorResult,
 } from "../../../utilities";
 import {
@@ -19,7 +18,7 @@ import {
 } from "../../../GenericFunctions";
 import type { IOperationHandler } from "../../../types/operation.types";
 
-import { DEBUG } from "../../../debug";
+import { DEBUG, debugLog } from "../../../debug";
 
 /**
  * Implementation of the "Get Many Records" operation
