+++ src/nodes/SurrealDb/resources/table/operations/listTables.operation.ts
@@ -10,10 +10,10 @@ import {
     buildCredentialsObject,
     checkQueryResult,
 } from "../../../GenericFunctions";
-import { debugLog, addErrorResult } from "../../../utilities";
+import { addErrorResult  } from "../../../utilities"
 import type { IOperationHandler } from "../../../types/operation.types";
 
-import { DEBUG } from "../../../debug";
+import { DEBUG, debugLog } from "../../../debug";
 
 /**
  * Implementation of the "List Tables" operation
