+++ src/nodes/SurrealDb/resources/table/operations/mergeAllRecords.operation.ts
@@ -7,7 +7,6 @@ import { NodeOperationError } from "n8n-workflow";
 import type { Surreal } from "surrealdb";
 import {
     formatArrayResult,
-    debugLog,
     createErrorResult,
 } from "../../../utilities";
 import {
@@ -21,7 +20,7 @@ import {
 } from "../../../GenericFunctions";
 import type { IOperationHandler } from "../../../types/operation.types";
 
-import { DEBUG } from "../../../debug";
+import { DEBUG, debugLog } from "../../../debug";
 
 /**
  * Merge All Records operation handler for Table resource
